# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface FplGameweek {
  id: number;
  name: string;
  deadlineTime: string;
  finished: boolean;
  isCurrent: boolean;
  isNext: boolean;
  averageScore?: number;
  highestScore?: number;
}

export interface FplPlayer {
  id: number;
  firstName: string;
  secondName: string;
  webName: string;
  position: string;
  team: number;
  teamName: string;
  price: number;
  totalPoints: number;
  eventPoints: number;
  form: number;
  selectedByPercent: number;
  transfersInEvent: number;
  transfersOutEvent: number;
  pointsPerGame?: number;
  minutes?: number;
  status: string;
  news: string;
  photo: string;
  priceRiseChance?: number;
  priceRiseLabel?: string;
  momentum?: string;
}

export interface FplOverview {
  currentGameweek: FplGameweek;
  nextGameweek: FplGameweek;
  players: FplPlayer[];
  updatedAt: string;
}

export interface FplTeam {
  id: number;
  name: string;
  managerName: string;
  overallRank: number;
  totalPoints: number;
  gameweekPoints: number;
  gameweekRank: number;
  bank: number;
  teamValue: number;
  transfersMade: number;
  lastDeadline: string;
}

export interface FplPick {
  playerId: number;
  webName: string;
  position: string;
  teamName: string;
  price: number;
  points: number;
  multiplier: number;
  isCaptain: boolean;
  isViceCaptain: boolean;
  purchasePrice: number;
  sellingPrice: number;
  fixture: string;
}

export interface FplTeamPicks {
  teamId: number;
  event: number;
  points: number;
  picks: FplPick[];
  updatedAt: string;
}

export interface FplRecommendation {
  playerId: number;
  webName: string;
  position: string;
  teamName: string;
  price: number;
  score: number;
  reason: string;
  priceRiseChance: number;
  fixture: string;
  replaces?: string;
}

export interface FplRecommendations {
  teamId: number;
  freeTransfers: number;
  recommendations: FplRecommendation[];
  watchlist: FplPlayer[];
  updatedAt: string;
}