# lib/api-zod/src/generated/types/fplOverview.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { FplGameweek } from './fplGameweek';
import type { FplPlayer } from './fplPlayer';

export interface FplOverview {
  currentGameweek: FplGameweek;
  nextGameweek: FplGameweek;
  players: FplPlayer[];
  updatedAt: string;
}