# lib/api-zod/src/generated/types/fplPick.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface FplPick {
  playerId: number;
  webName: string;
  position: string;
  teamName: string;
  price: number;
  points: number;
  multiplier: number;
  isCaptain: boolean;
  isViceCaptain: boolean;
  purchasePrice: number;
  sellingPrice: number;
  fixture: string;
}