# lib/api-zod/src/generated/types/fplPlayer.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface FplPlayer {
  id: number;
  firstName: string;
  secondName: string;
  webName: string;
  position: string;
  team: number;
  teamName: string;
  price: number;
  totalPoints: number;
  eventPoints: number;
  form: number;
  selectedByPercent: number;
  transfersInEvent: number;
  transfersOutEvent: number;
  pointsPerGame?: number;
  minutes?: number;
  status: string;
  news: string;
  photo: string;
  priceRiseChance?: number;
  priceRiseLabel?: string;
  momentum?: string;
}