# lib/api-zod/src/generated/types/fplRecommendation.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface FplRecommendation {
  playerId: number;
  webName: string;
  position: string;
  teamName: string;
  price: number;
  score: number;
  reason: string;
  priceRiseChance: number;
  fixture: string;
  replaces?: string;
}