# lib/api-zod/src/generated/types/fplRecommendations.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { FplPlayer } from './fplPlayer';
import type { FplRecommendation } from './fplRecommendation';

export interface FplRecommendations {
  teamId: number;
  freeTransfers: number;
  recommendations: FplRecommendation[];
  watchlist: FplPlayer[];
  updatedAt: string;
}