# lib/api-zod/src/generated/types/fplTeam.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface FplTeam {
  id: number;
  name: string;
  managerName: string;
  overallRank: number;
  totalPoints: number;
  gameweekPoints: number;
  gameweekRank: number;
  bank: number;
  teamValue: number;
  transfersMade: number;
  lastDeadline: string;
}