# lib/api-zod/src/generated/types/fplTeamPicks.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { FplPick } from './fplPick';

export interface FplTeamPicks {
  teamId: number;
  event: number;
  points: number;
  picks: FplPick[];
  updatedAt: string;
}